requirements.txt
aioredis==1.3.1
asgiref==3.2.7
asn1crypto==0.24.0
async-timeout==3.0.1
asyncore-wsgi==0.0.9
attrs==19.3.0
autobahn==20.1.3
Automat==0.8.0
autopep8==1.5.2
bottle==0.12.17
CacheControl==0.12.6
cachetools==4.0.0
certifi==2019.6.16
cffi==1.12.3
channels==2.4.0
channels-redis==2.4.1
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.7
daphne==2.4.1
defusedxml==0.6.0
Django==3.0.5
django-allauth==0.41.0
django-azure-ad-auth==1.3.1
django-cors-headers==3.2.1
django-filter==2.2.0
django-friendship==1.8.2
django-idmap==1.0.3
django-language-field==0.0.3
django-model-utils==3.2.0
django-notifications-hq==1.5.0
django-popup-forms==1.0.3
django-rest-auth==0.9.5
django-rest-multiple-models==2.1.3
django-rest-swagger==2.2.0
django-silk==4.0.1
django-sslserver==0.22
djangorestframework==3.11.0
firebase-admin==3.2.1
google-api-core==1.16.0
google-api-python-client==1.7.11
google-auth==1.11.0
google-auth-httplib2==0.0.3
google-cloud-core==1.2.0
google-cloud-firestore==1.6.1
google-cloud-storage==1.25.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
gprof2dot==2019.11.30
grpcio==1.26.0
gunicorn==19.9.0
hiredis==1.0.1
httpie==1.0.2
httplib2==0.17.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
itypes==1.1.0
Jinja2==2.10.1
jsonfield==2.0.2
lxml==4.4.1
Markdown==3.1.1
MarkupSafe==1.1.1
msgpack==0.6.2
numpy==1.17.0
oauthlib==3.1.0
openapi-codec==1.3.2
pandas==0.25.0
Pillow==7.1.1
protobuf==3.11.2
psycopg2-binary==2.8.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.19
Pygments==2.4.2
PyHamcrest==2.0.0
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
service-identity==18.1.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
Twisted==19.10.0
txaio==20.1.1
uritemplate==3.0.0
urllib3==1.25.3
web-pdb==1.5.1
wfastcgi==3.0.0
whitenoise==4.1.4
zope.interface==4.7.1
